server-a-words/requirements.txt
grpcio==1.65.1
protobuf==4.25.3